Extract sound functions into a module.
